test(api/v1alpha1): centralize api test helper
