+++ api/v1alpha1/databackup_types_test.go
@@ -60,10 +60,18 @@ var _ = Describe("DataBackup types", func() {
 			Expect(ok).To(BeTrue())
 			Expect(copiedDataBackup).NotTo(BeIdenticalTo(dataBackup))
 			Expect(copiedDataBackup.Spec).To(Equal(dataBackup.Spec))
-			Expect(copiedDataBackup.Spec.RunAs).NotTo(BeIdenticalTo(dataBackup.Spec.RunAs))
-			Expect(copiedDataBackup.Spec.RunAs.UID).NotTo(BeIdenticalTo(dataBackup.Spec.RunAs.UID))
-			Expect(copiedDataBackup.Spec.RunAs.GID).NotTo(BeIdenticalTo(dataBackup.Spec.RunAs.GID))
-			Expect(copiedDataBackup.Spec.RunAfter).NotTo(BeIdenticalTo(dataBackup.Spec.RunAfter))
+			if dataBackup.Spec.RunAs != nil {
+				Expect(copiedDataBackup.Spec.RunAs).NotTo(BeIdenticalTo(dataBackup.Spec.RunAs))
+				if dataBackup.Spec.RunAs.UID != nil {
+					Expect(copiedDataBackup.Spec.RunAs.UID).NotTo(BeIdenticalTo(dataBackup.Spec.RunAs.UID))
+				}
+				if dataBackup.Spec.RunAs.GID != nil {
+					Expect(copiedDataBackup.Spec.RunAs.GID).NotTo(BeIdenticalTo(dataBackup.Spec.RunAs.GID))
+				}
+			}
+			if dataBackup.Spec.RunAfter != nil {
+				Expect(copiedDataBackup.Spec.RunAfter).NotTo(BeIdenticalTo(dataBackup.Spec.RunAfter))
+			}
 			Expect(copiedDataBackup.Status).To(Equal(dataBackup.Status))
 
 			dataBackupList := &DataBackupList{Items: []DataBackup{*dataBackup}}
